+++ src/index.d.ts
@@ -67,6 +67,9 @@ export interface SchemaJSON {
     required?: Array<string>;
     compoundIndexes?: Array<string | Array<string>>;
     disableKeyCompression?: boolean;
+    attachments?: {
+            encrypted?: boolean
+    };
 }
 
 /**
@@ -95,6 +98,9 @@ export interface RxCollectionCreator {
         [key: string]: Function
     };
     methods?: {
+        [key: string]: Function
+    };
+    attachments?: {
         [key: string]: Function
     };
 }
@@ -161,6 +167,19 @@ export interface SyncOptions {
     // for options see https://pouchdb.com/api.html#replication
     options?: PouchReplicationOptions,
     query?: RxQuery<any>
+}
+
+export declare class RxAttachment<RxDocumentType> {
+    doc: RxDocument<RxDocumentType>;
+    id: string;
+    type: string;
+    length: number;
+    digest: string;
+    rev: string;
+
+    remove(): Promise<void>;
+    getData(): Promise<Blob>;
+    getStringData(): Promise<string>;
 }
 
 export type RxCollectionHookCallback<RxDocumentType> = (doc: RxDocument<RxDocumentType>) => void;
@@ -290,6 +309,11 @@ export declare class RxDocumentBase<RxDocumentType> {
     populate(objPath: string): Promise<RxDocument<RxDocumentType> | any>;
     update(updateObj: any): Promise<any>;
     atomicUpdate(fun: Function): Promise<RxDocument<RxDocumentType>>;
+
+    putAttachment(id: string, data: string, type?: string): Promise<RxAttachment<RxDocumentType>>;
+    getAttachment(id: string): Promise<RxAttachment<RxDocumentType>>;
+    allAttachments(): Promise<RxAttachment<RxDocumentType>[]>;
+    allAttachments$: Observable<RxAttachment<RxDocumentType>[]>;
 
     toJSON(): RxDocumentType;
     destroy(): void;